Critical flaw found and fixed + content
While developing the process for the playRound and plotReader, somehow we forgot that the game being displayed by the plotReader is not a 1 to 1 with the for-loop in playRound. By the time a script in plotReader is called to change, say, a condition of one of the rats, the full round has already been completed, rendering the changed condition useless for the games process. A Rat who's turn comes first might 'kill' another rat who comes last, but that rat wont officially be dead until the script is called way later in plotReader. The 'dead' rat will therefore still be 'alive' to play its turn during the actual game.

Big mistake...We concluded that the easiest fix would be to separate the background game sequence from the display by creating a 1 to 1 copy of the global rats[{}] obj array whose objects will be represented in the display instead. This of course required some additions for: addRat, randomRatBtn, the kill switches for the addToRatDisplay(), etc.. anything related to the creation of the rats array had to involve the new ratsDis[{}] (rats Display) object array. We managed to make a function, copyRat(), that copies the values for the properties of the original rats so as to avoid referencing as well. 

When events occur in the game, whatever conditions are changed for objects in rats[{}] is mirrored in ratsDis[{}], changes in ratsDis are recorded and evaluated later for the display in plotReader. We've essentially created a perfect 'replay' of background game.

also to note:
- the creation of antagRat in the events requires a copy as well. The new property, 'num', in the rats[{}] objects is crucial for referencing the the correct integer for the obj in ratsDis[{}]