events.moving update
In addition to rats freely moving from location to location, there is a chance they may encounter a trap or an enemy!

Traps require high 'wit' to be bested, enemies require high 'ferocity', the reward is food. failure in the encounter results in being wounded, and if already wounded, death. 

Wounded rats have a yellow display now.