scroller and chance changes
Changed foodChance to statChance that takes in the int value of any protagRat stat and creates an array of 0's and 1's. The number of 1's correspond with the stat given, as well as the 0's made from the difference of 10 (10 - stat). function returns a random value in the array, true (1) or false (0). The higher the stat, the better chance of returning true.
Some event moves that rely on the stat of a rat now use statChance to determine their success.

added scrolling capabilities to text-box after a round is finished