Desperation night event
Added the desperation() night event for rats that end up starved going into nightfall. 

Currently there is a chance for rats to encounter the spider, resulting in a hungry stomach or death, OR scraps, keeping them alive for the night.

Rats should not be able to pass to the next round if starving. If food is unavailable, death is certain. 

-later add encounter for other rats in the same locale.