Room Destruction - complete fix
Realized that I fell into the same trap as before with seeing ratsDis[{}] and rats[{}] as simultaneous when the background game is working. Rat movement from one room to the next caused a crash because I forgot ratsDis does not update stats at the same time, only later when the game is displayed. I could not call a ratsDis{} from the same room as its corresponding background rat{} after that rat{} has already changed rooms in the 'background' game. largely the issue stemmed in collapse() simple fix by utilizing the rats{}.num property (which references its integer in the array) to push the needed ratsDis{} into the new array needed for escapeTurn(). 

Also solved the supposed issue with escapeTurnP2(), the display txt for being crushed was the same for being wounded...whoopsie. 

~note to self to take a break when stuck. go to sleep after 11pm etc.