Winner and Restart + html
forgot the html file update for the Winner and Restart Commit, whoopsie.

just another button within the game-container div. The round/proceed  button is switched with the restart  button once a winner is found and the game is done. Simple use of style.display = 'none' to remove and insert it as needed.